move getPage to its own file
